requirements.txt
Flask
redis
psycopg2-binary
python-dotenv
